requirements.txt
# UnifyInbox - Requirements
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1

# AI
anthropic==0.40.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Cache & Queue
redis==5.2.0
celery==5.4.0

# Vector Store
chromadb==0.5.20

# HTTP Client
httpx==0.28.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==44.0.0

# Google (Gmail)
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.157.0

# Slack
slack-sdk==3.33.4

# Telegram
python-telegram-bot==21.9

# Discord / WebSocket
websockets==14.1

# Utilities
python-multipart==0.0.12
email-validator==2.2.0
